lint fixes for severity to float

+++ Packs/MicrosoftTeams/Integrations/MicrosoftTeams/MicrosoftTeams.py
@@ -1,6 +1,6 @@
 import demistomock as demisto
 from CommonServerPython import *
-# from CommonServerUserPython import *
+from CommonServerUserPython import *
 
 ''' IMPORTS '''
 import requests
@@ -85,21 +85,21 @@ def error_parser(resp_err: requests.Response, api: str = 'graph') -> str:
         return resp_err.text
 
 
-def translate_severity(severity: str) -> int:
+def translate_severity(severity: str) -> float:
     """
     Translates Demisto text severity to int severity
     :param severity: Demisto text severity
     :return: Demisto integer severity
     """
     severity_dictionary = {
-        'Unknown': 0,
+        'Unknown': 0.0,
         'Informational': 0.5,
-        'Low': 1,
-        'Medium': 2,
-        'High': 3,
-        'Critical': 4
+        'Low': 1.0,
+        'Medium': 2.0,
+        'High': 3.0,
+        'Critical': 4.0
     }
-    return severity_dictionary.get(severity, 0)
+    return severity_dictionary.get(severity, 0.0)
 
 
 def create_incidents(demisto_user: dict, incidents: list) -> dict:
@@ -1162,7 +1162,7 @@ def send_message():
             or channel_name == INCIDENT_NOTIFICATIONS_CHANNEL:
         # Got a notification from server
         channel_name = demisto.params().get('incident_notifications_channel', 'General')
-        severity: int = int(demisto.args().get('severity'))
+        severity: float = float(demisto.args().get('severity'))
 
         # Adding disable and not enable because of adding new boolean parameter always defaults to false value in server
         if (disable_auto_notifications := demisto.params().get('disable_auto_notifications')) is not None:
@@ -1171,7 +1171,7 @@ def send_message():
             disable_auto_notifications = False
 
         if not disable_auto_notifications:
-            severity_threshold: int = translate_severity(demisto.params().get('min_incident_severity', 'Low'))
+            severity_threshold: float = translate_severity(demisto.params().get('min_incident_severity', 'Low'))
             if severity < severity_threshold:
                 return
         else:
